% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCIheatmap.R
\name{CCIheatmap}
\alias{CCIheatmap}
\title{plot heatmap of cell-cell interaction scores for all LR pairs}
\usage{
CCIheatmap(CCI_df, include_celltypes = NULL)
}
\arguments{
\item{CCI_df}{dataframe of cell-cell interaction scores from CCI anlaysis}
}
\description{
plot heatmap of cell-cell interaction scores for all LR pairs
}
