% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCIspatial.R
\name{CCIspatial}
\alias{CCIspatial}
\title{plot spatial plot of cell-cell interaction pattern for a specific LR pair}
\usage{
CCIspatial(
  spe,
  BLISA_output,
  index,
  ct_group = "cell_type",
  top = 30,
  hex_size = 50,
  dmax = 250
)
}
\arguments{
\item{spe}{cell-level spe object}

\item{BLISA_output}{result outout from runBLISA}

\item{index}{LR index}
}
\description{
plot spatial plot of cell-cell interaction pattern for a specific LR pair
}
