% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blisa-class.R
\name{is.blisa}
\alias{is.blisa}
\title{Test if an object is a blisa object}
\usage{
is.blisa(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical.
}
\description{
Test if an object is a blisa object
}
\examples{
is.blisa(list())           # FALSE
is.blisa("not a blisa")    # FALSE
}
