% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLRIsum.R
\name{plotLRIsum}
\alias{plotLRIsum}
\title{plot dot plot ranking LR results from BLISA}
\usage{
plotLRIsum(LR_out, top = NULL, pt.size = 4)
}
\arguments{
\item{LR_out}{BLISA LRI result}
}
\description{
plot dot plot ranking LR results from BLISA
}
