% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBLISA.R
\name{runBLISA.old}
\alias{runBLISA.old}
\title{Run BLISA analysis}
\usage{
runBLISA.old(
  coords_df,
  counts_matrix,
  LR_df,
  hex_size = 50,
  dmax = 250,
  nsim = 999,
  p_cutoff = 0.05,
  min_ligand = 10,
  min_receptor = 10,
  col = "annotation",
  default_mode = "diffuse",
  diffuse_category = c("Secreted Signaling", "Non-protein Signaling")
)
}
\arguments{
\item{coords_df}{spatial coordimates dataframe}

\item{counts_matrix}{Gene expression matrix}

\item{LR_df}{Ligand-receptor dataframe}
}
\description{
Run BLISA analysis
}
