% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBLISA_spe.R
\name{runBLISA.spe}
\alias{runBLISA.spe}
\title{Run BLISA analysis on cell-level spe}
\usage{
runBLISA.spe(
  spe,
  sf = NULL,
  LR_df,
  hex_size = 50,
  dmax = 250,
  nsim = 999,
  p_cutoff = 0.05,
  min_ligand = 10,
  min_receptor = 10,
  col = "annotation",
  default_mode = "diffuse",
  diffuse_category = c("Secreted Signaling", "Non-protein Signaling")
)
}
\arguments{
\item{spe}{cell-level spe object}

\item{LR_df}{Ligand-receptor dataframe}

\item{bin_sf}{Bin-level sf object}
}
\description{
Run BLISA analysis on cell-level spe
}
