% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{blisa-package}
\alias{blisa-package}
\title{blisa: Bivariate Local Indicator of Spatial Association for spatial transcriptomics}
\description{
Identifies cell-cell communication hotspots in spatial transcriptomics data using bivariate Local Moran's I statistics on hexagonally binned cells. Provides functions for spatial weighting, ligand-receptor pair filtering, hotspot detection, and visualisation of sender-receiver cell-type interactions.
}
\author{
\strong{Maintainer}: Yunshun Chen \email{yuchen@wehi.edu.au}

Authors:
\itemize{
  \item Lei Qin \email{qin.l@wehi.edu.au}
  \item Lizhong Chen
}

}
