% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCIheatmapOneLR.R
\name{CCIheatmapOneLR}
\alias{CCIheatmapOneLR}
\title{CCI heatmap for one ligand-receptor pair
...}
\usage{
CCIheatmapOneLR(CCIres, lr_pair)
}
\description{
CCI heatmap for one ligand-receptor pair
...
}
