% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLRpairs.R
\name{filterLRpairs}
\alias{filterLRpairs}
\title{Filter ligand-receptor pairs
...}
\usage{
filterLRpairs(
  counts,
  min_ligand = 10,
  min_receptor = 10,
  LR_df = CellChatDB.human$interaction
)
}
\description{
Filter ligand-receptor pairs
...
}
