% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLRI.R
\name{plotLRI.sf}
\alias{plotLRI.sf}
\title{Run BLISA analysis on bin level sf}
\usage{
plotLRI.sf(hex_sf, LR_out, index, log_pval = TRUE)
}
\arguments{
\item{hex_sf}{Bin-level sf object}

\item{LR_out}{BLISA LRI result}

\item{index}{Ligand-receptor index}
}
\description{
Run BLISA analysis on bin level sf
}
