% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBLISA_default.R
\name{runBLISA.default}
\alias{runBLISA.default}
\title{Run BLISA analysis on bin level sf}
\usage{
runBLISA.default(
  counts_matrix,
  bin_sf,
  LR_df,
  hex_size = 50,
  dmax = 250,
  nsim = 999,
  p_cutoff = 0.05,
  min_ligand = 10,
  min_receptor = 10,
  col = "annotation",
  default_mode = "diffuse",
  diffuse_category = c("Secreted Signaling", "Non-protein Signaling")
)
}
\arguments{
\item{counts_matrix}{Gene expression matrix}

\item{bin_sf}{Bin-level sf object}

\item{LR_df}{Ligand-receptor dataframe}
}
\description{
Run BLISA analysis on bin level sf
}
