% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBLISA_spe.R
\name{runBLISA.spe.isolates.removed}
\alias{runBLISA.spe.isolates.removed}
\title{Run BLISA on Bin-Level Spatial Data with Isolate Removal}
\usage{
runBLISA.spe.isolates.removed(
  spe,
  sf = NULL,
  LR_df,
  hex_size = 50,
  dmax = 250,
  nsim = 999,
  p_cutoff = 0.05,
  min_ligand = 10,
  min_receptor = 10,
  min_cells_per_bin = 1,
  col = "annotation",
  default_mode = "diffuse",
  diffuse_category = c("Secreted Signaling", "Non-protein Signaling")
)
}
\arguments{
\item{min_cells_per_bin}{Integer. Minimum number of cells a hex bin must
contain to be included in spatial weight computation and Moran's I
calculation. Bins below this threshold are excluded and assigned neutral
statistics (p = 1, LISA = 0).}
}
\description{
Run BLISA on Bin-Level Spatial Data with Isolate Removal
}
