% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCCI.R
\name{runCCI}
\alias{runCCI}
\title{run CCI analysis to get cell-cell interaction scores for all cell type pairs and ligand-receptor pairs
...}
\usage{
runCCI(spe, BLISA_output, ct_group = "cell_type", hex_size = 50, dmax = 250)
}
\description{
run CCI analysis to get cell-cell interaction scores for all cell type pairs and ligand-receptor pairs
...
}
